+++ Runtime/SDK/Android/SDK/ApphudFacebookAttribution.cs
@@ -1,6 +1,7 @@
 #if UNITY_ANDROID && APPHUD_FB
 using System;
 using System.Collections.Generic;
+using System.Text;
 using Apphud.Unity.Domain;
 using UnityEngine;
 
@@ -12,7 +13,7 @@ namespace Apphud.Unity.Android.SDK
         private const string AndroidAppEventsLoggerClass = "com.facebook.appevents.AppEventsLogger";
         private const string AndroidJSONObjectClass = "org.json.JSONObject";
 
-        internal static void Add()
+        internal static void Add(Action<string> onError = null)
         {
             using (AndroidJavaClass unityPlayer = new AndroidJavaClass("com.unity3d.player.UnityPlayer"))
             {
@@ -20,17 +21,42 @@ namespace Apphud.Unity.Android.SDK
 
                 using (AndroidJavaObject jsonObject = new AndroidJavaObject(AndroidJSONObjectClass))
                 {
-                    using (AndroidJavaClass utilityClass = new AndroidJavaClass(AndroidUtilityClass))
+                    StringBuilder errorBuilder = new StringBuilder();
+
+                    string extInfo = "";
+                    try
                     {
-                        utilityClass.CallStatic("setAppEventExtendedDeviceInfoParameters", jsonObject, activity);
+                        using (AndroidJavaClass utilityClass = new AndroidJavaClass(AndroidUtilityClass))
+                        {
+                            utilityClass.CallStatic("setAppEventExtendedDeviceInfoParameters", jsonObject, activity);
+                        }
+
+                        extInfo = jsonObject.Call<string>("get", "extinfo");
+                    }
+                    catch (Exception ex)
+                    {
+                        errorBuilder.AppendLine($"Error on getting extinfo: {ex.Message}");
                     }
 
-                    string extInfo = jsonObject.Call<string>("get", "extinfo");
-
-                    string anonID;
-                    using (AndroidJavaClass appEventsLoggerClass = new AndroidJavaClass(AndroidAppEventsLoggerClass))
+                    string anonID = "";
+                    try
                     {
-                        anonID = appEventsLoggerClass.CallStatic<string>("getAnonymousAppDeviceGUID", activity);
+                        using (AndroidJavaClass appEventsLoggerClass = new AndroidJavaClass(AndroidAppEventsLoggerClass))
+                        {
+                            anonID = appEventsLoggerClass.CallStatic<string>("getAnonymousAppDeviceGUID", activity);
+                        }
+                    }
+                    catch (Exception ex)
+                    {
+                        errorBuilder.AppendLine($"Error on getting anonID: {ex.Message}");
+                    }
+
+                    string error = errorBuilder.ToString();
+
+                    if (error.Length > 0)
+                    {
+                        onError?.Invoke(error);
+                        return;
                     }
 
                     Dictionary<string, object> attributionData = new Dictionary<string, object>